backend/requirements.txt
# Core Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
typing-extensions==4.12.2  # Required for typing.Literal and other type hints

# Database
pymongo==4.9.0
motor==3.6.0

# AI/ML Services
google-generativeai==0.8.0
groq==0.32.0
openai==1.51.0
chromadb==0.4.24
sentence-transformers==3.1.0

# OCR & Document Processing
easyocr==1.7.2
pytesseract==0.3.13
pillow==10.4.0
pdf2image==1.17.0
opencv-python-headless==4.9.0.80  # Required for EasyOCR
torch==2.2.2  # Required for EasyOCR
torchvision==0.17.2  # Required for image processing

# Machine Learning
xgboost==2.1.1
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.4

# Voice/Audio Processing
openai-whisper==20231117
ffmpeg-python==0.2.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.0
playwright==1.47.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
